Add Ui interface and refactor existing Ui to Tui
