+++ src/main/java/tohru/Tohru.java
@@ -6,6 +6,7 @@ import tohru.exception.CorruptSaveException;
 import tohru.exception.TohruException;
 import tohru.storage.FileStore;
 import tohru.task.TodoList;
+import tohru.ui.Tui;
 import tohru.ui.Ui;
 
 /**
@@ -16,14 +17,15 @@ public class Tohru {
     private FileStore store;
     private TodoList todoList;
     private Ui ui;
+    private boolean isExit = false;
 
     /**
      * Setups up the chatbot.
      *
      * @param filePath The location to store the save file.
      */
-    public Tohru(String filePath) {
-        ui = new Ui();
+    public Tohru(String filePath, Ui ui) {
+        this.ui = ui;
         store = new FileStore(filePath);
         try {
             todoList = new TodoList(store.retrieveTodoList());
@@ -40,30 +42,36 @@ public class Tohru {
     }
 
     /**
-     * Executes the main functionalities of the chatbot.
+     * Executes the main functionalities of the text-based chatbot.
      */
-    public void run() {
+    public void runTui() {
         // Greetings
         ui.showWelcome();
 
-        boolean isExit = false;
         while (!isExit) {
-            try {
-                String prompt = ui.readCommand();
-                ui.showDivider();
-                Command cmd = Parser.parse(prompt);
-                cmd.execute(todoList, ui, store);
-                isExit = cmd.isExit();
-            } catch (TohruException e) {
-                ui.showError(e.getMessage());
-            } finally {
-                ui.showDivider();
-            }
+            process();
+        }
+    }
+
+    /**
+     * Handles the request and responses related functionalities of the text-based chatbot.
+     */
+    public void process() {
+        try {
+            String prompt = ui.readCommand();
+            ui.showDivider();
+            Command cmd = Parser.parse(prompt);
+            cmd.execute(todoList, ui, store);
+            isExit = cmd.isExit();
+        } catch (TohruException e) {
+            ui.showError(e.getMessage());
+        } finally {
+            ui.showDivider();
         }
     }
 
     public static void main(String[] args) {
-        new Tohru("./data/todosData.txt").run();
+        new Tohru("./data/todosData.txt", new Tui()).runTui();
     }
 
 }
