+++ src/test/java/tohru/command/AddDeadlineCommandTest.java
@@ -12,6 +12,7 @@ import tohru.exception.TohruException;
 import tohru.storage.FileStore;
 import tohru.storage.FileStoreStub;
 import tohru.task.TodoList;
+import tohru.ui.Tui;
 import tohru.ui.Ui;
 
 
@@ -23,7 +24,7 @@ public class AddDeadlineCommandTest {
 
     @BeforeAll
     static void initAll() {
-        ui = new Ui();
+        ui = new Tui();
         store = new FileStoreStub();
     }
